requirements.txt
websocket-client
requests
sqlobject
